requirements.txt
faiss_cpu==1.9.0.post1
faiss_gpu_cu12==1.9.0.0
numpy==2.1.3
openai==1.55.0
pandas==2.2.3
Pillow==11.0.0
protobuf==5.28.3
Requests==2.32.3
sentence_transformers==3.3.1
spacy==3.8.2
torch==2.5.0
tqdm==4.66.5
transformers==4.46.3
huggingface-hub==0.26.2
